requirements.txt
astroid==1.3.8
beautifulsoup4==4.4.1
colorama==0.3.3
coverage==4.0.1
dateutils==0.6.6
decorator==4.0.4
Django==1.8.5
django-ckeditor==5.0.2
django-reversion==1.9.3
future==0.15.2
ipython==3.2.1
itsdangerous==0.24
jdcal==1.0
jedi==0.9.0
logilab-common==1.0.2
MarkupSafe==0.23
mccabe==0.3.1
mysql-connector-python==2.1.3
nose==1.3.7
numpy==1.8.2
openpyxl==2.2.6
path.py==8.1.2
pep8==1.6.2
pilkit==1.1.13
Pillow==3.0.0
pluggy==0.3.1
psutil==3.2.2
py==1.4.30
pyflakes==1.0.0
pylint==1.4.4
PyMySQL==0.6.7
pyparsing==2.0.3
pyquery==1.2.9
python-dateutil==2.4.2
pytz==2015.4
pywin32==219
qrcode==5.1
requests==2.8.0
selenium==2.48.0
six==1.10.0
sorl-thumbnail==12.3
SQLAlchemy==1.0.8
sqlparse==0.1.16
tox==2.1.1
virtualenv==13.1.2
Werkzeug==0.10.4
xlrd==0.9.4
XlsxWriter==0.7.6
